implemented draft version of adding a new company

+++ infrastructure/database/migration/20200303120519_crate_company_table.sql
@@ -1,7 +1,7 @@
 -- +goose Up
 -- SQL in this section is executed when the migration is applied.
 CREATE TABLE companies (
-	id bigint NOT NULL,
+	id serial NOT NULL,
 	uuid char(36) NOT NULL,
 	name varchar(80) NOT NULL,
 	created_at timestamp NOT NULL
